# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  sylve_internal.APIResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  sylve_internal.APIResponse-array_sylve_internal_db_models_info_IODelay:
    properties:
      data:
        items:
          $ref: '#/definitions/sylve_internal_db_models_info.IODelay'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  sylve_internal.APIResponse-array_sylve_internal_db_models_info_Note:
    properties:
      data:
        items:
          $ref: '#/definitions/sylve_internal_db_models_info.Note'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  sylve_internal.APIResponse-array_sylve_internal_interfaces_services_zfs_Zpool:
    properties:
      data:
        items:
          $ref: '#/definitions/sylve_internal_interfaces_services_zfs.Zpool'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  sylve_internal.APIResponse-float64:
    properties:
      data:
        type: number
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  sylve_internal.APIResponse-sylve_internal_interfaces_services_info_RAMInfo:
    properties:
      data:
        $ref: '#/definitions/sylve_internal_interfaces_services_info.RAMInfo'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  sylve_internal.APIResponse-sylve_internal_interfaces_services_info_SwapInfo:
    properties:
      data:
        $ref: '#/definitions/sylve_internal_interfaces_services_info.SwapInfo'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  sylve_internal_db_models_info.IODelay:
    properties:
      createdAt:
        type: string
      delay:
        type: number
      id:
        type: integer
    type: object
  sylve_internal_db_models_info.Note:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      updateAt:
        type: string
    type: object
...
  sylve_internal_interfaces_services_info.RAMInfo:
    properties:
      free:
        type: integer
      total:
        type: integer
      usedPercent:
        type: number
    type: object
  sylve_internal_interfaces_services_info.SwapInfo:
    properties:
      free:
        type: integer
      total:
        type: integer
      usedPercent:
        type: number
    type: object
  sylve_internal_interfaces_services_zfs.Zpool:
    properties:
      allocated:
        type: integer
      dedupRatio:
        type: number
      free:
        type: integer
      freeing:
        type: integer
      health:
        type: string
      leaked:
        type: integer
      name:
        type: string
      readOnly:
        type: boolean
      size:
        type: integer
    type: object
host: sylve.lan:8181
info:
  contact:
    email: hello@alchemilla.io
    name: Alchemilla Ventures Pvt. Ltd.
    url: https://alchemilla.io
  description: Sylve is a lightweight GUI for managing Bhyve, Jails, ZFS, networking,
    and more on FreeBSD.
  license:
    name: BSD-2-Clause
    url: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  termsOfService: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  title: Sylve API
  version: 0.0.1
paths:
...
  /info/notes:
    get:
      consumes:
      - application/json
      description: Get all notes stored in the database
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-array_sylve_internal_db_models_info_Note'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get All Notes
      tags:
      - Info
    post:
      consumes:
      - application/json
      description: Add a new note to the database
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - Info
  /info/notes/:id:
    delete:
      consumes:
      - application/json
      description: Delete a note from the database by its ID
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "400":
          description: Invalid note ID
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete a note by ID
      tags:
      - Info
    put:
      consumes:
      - application/json
      description: Update a note in the database by its ID
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "400":
          description: Invalid note ID
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Update a note by ID
      tags:
      - Info
  /info/ram:
    get:
      consumes:
      - application/json
      description: Get the RAM information about the system
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-sylve_internal_interfaces_services_info_RAMInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get RAM Info
      tags:
      - Info
  /info/swap:
    get:
      consumes:
      - application/json
      description: Get the Swap information about the system
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-sylve_internal_interfaces_services_info_SwapInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Swap Info
      tags:
      - Info
  /zfs/avg-io-delay:
    get:
      consumes:
      - application/json
      description: Get the average IO delay of all pools
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-float64'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Average IO Delay
      tags:
      - ZFS
  /zfs/avg-io-delay-historical:
    get:
      consumes:
      - application/json
      description: Get the historical IO delays of all pools
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-array_sylve_internal_db_models_info_IODelay'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Average IO Delay Historical
      tags:
      - ZFS
  /zfs/pools:
    get:
      consumes:
      - application/json
      description: Get all ZFS pools
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-array_sylve_internal_interfaces_services_zfs_Zpool'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Pools
      tags:
      - ZFS
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"